Exclude wordpress-plugin installation paths
When modules depend on packages of type `wordpress-plugin` we don't want to install them to `wp-content/plugins`, but instead just in the `vendor` directory like any typical package. To do this we have to override the Composer Installers class to a patched version that excludes specific plugins from going to their normal location.

Ideally we'd have the list of excluded plugins be pulled from module's `composer.json` files, so we don't need to maintain a centralized list. The problem is I can't find a way get all the cached `package.json` content from the installer location. I'm pretty sure this _is_ in memory at the point of installing any `wordpress-plugin`, but that assumption may not be right. I have a feeling the complete dep graph isn't available until after the Install Manager installs all the packages.

For now, I think we should continue with a hardcoded list, and when someone wants to have a stab at upgrading this to support specifying excluded plugins from a module's `composer.json` then we add support for that.