Define WP_ENVIRONMENT_TYPE
As part of the WP 5.5 compatibility work this mirrors the `HM_ENV_TYPE` constant to the new core `WP_ENVIRONMENT_TYPE` constant which defaults to "local". This will seemlessly work with the cloud module `wp-config.php` and `wp-config-produciton.php` files should we add `WP_ENVIRONMENT_TYPE` as an environment variable to our infrastructure in future.